app/requirements.txt
flask
psycopg2-binary
flask-cors
flask-sqlalchemy
flask-migrate
PyJWT
pytest
redis
